Move process stuff to utils/processes
