+++ src/commands/open_in_other_editors.ts
@@ -1,10 +1,11 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as vs from "vscode";
-import { isMac, safeSpawn } from "../debug/utils";
+import { isMac } from "../debug/utils";
 import { androidStudioPath, flutterPath } from "../sdk/utils";
 import { fsPath, Sdks } from "../utils";
 import { logError } from "../utils/log";
+import { safeSpawn } from "../utils/processes";
 
 export class OpenInOtherEditorCommands implements vs.Disposable {
 	private disposables: vs.Disposable[] = [];
