+++ src/commands/sdk.ts
@@ -7,7 +7,7 @@ import * as vs from "vscode";
 import { ProgressLocation, Uri, window } from "vscode";
 import { config } from "../config";
 import { stripMarkdown } from "../dartdocs";
-import { flatMap, globalFlutterArgs, LogCategory, LogSeverity, PromiseCompleter, safeSpawn } from "../debug/utils";
+import { flatMap, LogCategory, LogSeverity, PromiseCompleter } from "../debug/utils";
 import { FlutterCapabilities } from "../flutter/capabilities";
 import { FlutterDeviceManager } from "../flutter/device_manager";
 import { locateBestProjectRoot } from "../project";
@@ -21,8 +21,8 @@ import * as util from "../utils";
 import { fsPath, ProjectType, Sdks } from "../utils";
 import { sortBy } from "../utils/array";
 import { getChildFolders, hasPubspec } from "../utils/fs";
-import { log } from "../utils/log";
-import { logProcess } from "../utils/processes";
+import { log, logProcess } from "../utils/log";
+import { globalFlutterArgs, safeSpawn } from "../utils/processes";
 import * as channels from "./channels";
 
 const packageNameRegex = new RegExp("^[a-z][a-z0-9_]*$");
