+++ src/sdk/dev_tools.ts
@@ -2,12 +2,12 @@ import * as child_process from "child_process";
 import * as path from "path";
 import * as vs from "vscode";
 import { Analytics } from "../analytics";
-import { LogCategory, LogSeverity, safeSpawn } from "../debug/utils";
+import { LogCategory, LogSeverity } from "../debug/utils";
 import { PubGlobal } from "../pub/global";
 import { openInBrowser, Sdks } from "../utils";
 import { DartDebugSessionInformation, extractObservatoryPort } from "../utils/debug";
-import { log, logError } from "../utils/log";
-import { logProcess } from "../utils/processes";
+import { log, logError, logProcess } from "../utils/log";
+import { safeSpawn } from "../utils/processes";
 import { pubPath } from "./utils";
 
 const devtools = "devtools";
